refactor(cli): split remaining run_* into packaging and transcribe modules
Move download/quantize/cache-gc into packaging.rs and the offline
transcribe/batch/watch path into transcribe_cmd.rs. Serve-only
build_limits/build_server_config live next to run_serve. main.rs
is the clap dispatcher.

Constraint: no CLI behavior change
Rejected: one giant cli.rs | keep command families in separate files
Confidence: high
Scope-risk: moderate
