+++ crates/gigastt/src/main.rs
@@ -1,19 +1,17 @@
-use anyhow::Context;
-use clap::{Args, CommandFactory, FromArgMatches, Parser, Subcommand};
-use gigastt::batch;
-use gigastt::boot::{
-    EngineRecipe, ItnMode, PunctuationMode, parse_itn_mode, parse_punctuation_mode,
-};
-use gigastt::server;
-use gigastt::server::{OriginPolicy, RuntimeLimits, ServerConfig};
-use gigastt_core::export::{ExportFormat, RenderOpts};
+use clap::{CommandFactory, FromArgMatches, Parser, Subcommand};
+use gigastt::boot::{ItnMode, PunctuationMode, parse_itn_mode, parse_punctuation_mode};
+use gigastt_core::model;
 use gigastt_core::model::{ModelVariant, ProgressMode};
-use gigastt_core::{inference, model};
-use std::str::FromStr;
 use tracing_subscriber::EnvFilter;
 
+mod packaging;
 mod serve;
+mod transcribe_cmd;
+use packaging::{run_cache_gc, run_download, run_quantize};
 use serve::{ServeArgs, run_serve};
+use transcribe_cmd::{
+    BatchOutputArgs, OfflineEngineArgs, run_transcribe, run_transcribe_batch, run_watch,
+};
 
 #[derive(Parser)]
 #[command(
@@ -35,140 +33,6 @@ struct Cli {
 
     #[command(subcommand)]
     command: Commands,
-}
-
-/// Engine / post-processing flags shared by the offline directory commands
-/// (`transcribe-batch`, `watch`). Mirrors the corresponding `transcribe` flags.
-#[derive(Args)]
-struct OfflineEngineArgs {
-    /// Model directory
-    #[arg(long, default_value_t = model::default_model_dir())]
-    model_dir: String,
-
-    /// Recognition head to use. Omit to auto-detect from the model directory.
-    /// Env: GIGASTT_MODEL_VARIANT.
-    #[arg(
-        long,
-        env = "GIGASTT_MODEL_VARIANT",
-        value_parser = parse_model_variant
-    )]
-    model_variant: Option<ModelVariant>,
-
-    /// Punctuation + capitalization restoration: `on`, `off`, or `auto`.
-    /// Env: GIGASTT_PUNCTUATION.
-    #[arg(
-        long,
-        env = "GIGASTT_PUNCTUATION",
-        default_value = "auto",
-        value_parser = parse_punctuation_mode
-    )]
-    punctuation: PunctuationMode,
-
-    /// Directory holding the optional punctuation model.
-    /// Env: GIGASTT_PUNCT_MODEL_DIR.
-    #[arg(
-        long,
-        env = "GIGASTT_PUNCT_MODEL_DIR",
-        default_value_t = model::default_punct_model_dir()
-    )]
-    punct_model_dir: String,
-
-    /// Inverse text normalization (Russian number-words → digits):
-    /// `on`, `off`, or `auto`. Env: GIGASTT_ITN.
-    #[arg(
-        long,
-        env = "GIGASTT_ITN",
-        default_value = "auto",
-        value_parser = parse_itn_mode
-    )]
-    itn: ItnMode,
-
-    /// Contextual hotword biasing: path to a file of phrases to boost (one
-    /// phrase per line, optional `\t<weight>` suffix). Env: GIGASTT_HOTWORDS_FILE.
-    #[arg(long, env = "GIGASTT_HOTWORDS_FILE")]
-    hotwords_file: Option<String>,
-
-    /// Also bias the built-in Russian brand/acronym lexicon.
-    /// Env: GIGASTT_HOTWORDS_DEFAULT.
-    #[arg(long, env = "GIGASTT_HOTWORDS_DEFAULT", default_value_t = false)]
-    hotwords_default: bool,
-
-    /// Additive logit boost applied to hotword continuation tokens [default: 5.0].
-    /// Env: GIGASTT_HOTWORDS_BOOST.
-    #[arg(long, env = "GIGASTT_HOTWORDS_BOOST")]
-    hotwords_boost: Option<f32>,
-
-    /// Voice activity detection: skip silence before decoding. Env: GIGASTT_VAD.
-    #[arg(long, env = "GIGASTT_VAD", default_value_t = false)]
-    vad: bool,
-
-    /// VAD speech-probability threshold in [0,1] [default: 0.5].
-    /// Env: GIGASTT_VAD_THRESHOLD.
-    #[arg(long, env = "GIGASTT_VAD_THRESHOLD")]
-    vad_threshold: Option<f32>,
-
-    /// Minimum trailing silence (ms) to close a speech region [default: 500].
-    /// Env: GIGASTT_VAD_MIN_SILENCE_MS.
-    #[arg(long, env = "GIGASTT_VAD_MIN_SILENCE_MS")]
-    vad_min_silence_ms: Option<u32>,
-
-    /// Directory holding the Silero VAD model (`silero_vad.onnx`).
-    /// Env: GIGASTT_VAD_MODEL_DIR.
-    #[arg(long, env = "GIGASTT_VAD_MODEL_DIR", default_value_t = model::default_vad_model_dir())]
-    vad_model_dir: String,
-
-    /// Intra-op thread count for the encoder session on the CPU build. When
-    /// unset, defaults to the logical CPU count divided across the pool.
-    /// Do not set `1` on multi-core hosts unless debugging — it is ~3× slower
-    /// than auto. Explicit `1` is still honoured for debug passthrough.
-    /// Env: GIGASTT_ENCODER_INTRA_THREADS.
-    #[arg(long, env = "GIGASTT_ENCODER_INTRA_THREADS")]
-    encoder_intra_threads: Option<usize>,
-
-    /// Number of concurrent transcription workers (engine session pool). Each
-    /// session loads its own encoder copy (~0.4 GB resident for the INT8
-    /// encoder). Default 2 suits multi-file hosts; use `--pool-size 1` on
-    /// edge / low-RAM (~400 MB RSS). Pool > 1 costs RAM and can cost ~10–20%
-    /// single-job RTF (threads split across slots).
-    #[arg(long, default_value_t = 2)]
-    pool_size: usize,
-}
-
-/// Output / source-policy flags shared by `transcribe-batch` and `watch`.
-#[derive(Args)]
-struct BatchOutputArgs {
-    /// Export formats, comma-separated: txt, json, md, srt, vtt.
-    /// One `<stem>.<ext>` file is written per format. Env: GIGASTT_FORMAT.
-    #[arg(short, long, env = "GIGASTT_FORMAT", default_value = "txt,json")]
-    format: String,
-
-    /// Move each successfully transcribed source file into this directory
-    /// (e.g. `--move-to done/`). Mutually exclusive with `--delete-source`.
-    /// Env: GIGASTT_BATCH_MOVE_TO.
-    #[arg(long, env = "GIGASTT_BATCH_MOVE_TO", conflicts_with = "delete_source")]
-    move_to: Option<String>,
-
-    /// Delete each successfully transcribed source file. Failed files are
-    /// always left in place. Env: GIGASTT_BATCH_DELETE_SOURCE.
-    #[arg(long, env = "GIGASTT_BATCH_DELETE_SOURCE", default_value_t = false)]
-    delete_source: bool,
-
-    /// Extra attempts per file after a failure, with a short backoff
-    /// [default: 0 for transcribe-batch, 2 for watch]. Env: GIGASTT_BATCH_RETRIES.
-    #[arg(long, env = "GIGASTT_BATCH_RETRIES")]
-    retries: Option<u32>,
-
-    /// Maximum characters per subtitle/caption line (SRT/VTT) [default: 80]
-    #[arg(long, env = "GIGASTT_MAX_CHARS_PER_LINE")]
-    max_chars_per_line: Option<usize>,
-
-    /// Maximum words per subtitle/caption line (SRT/VTT) [default: 14]
-    #[arg(long, env = "GIGASTT_MAX_WORDS_PER_LINE")]
-    max_words_per_line: Option<usize>,
-
-    /// Include per-word timestamps in Markdown output
-    #[arg(long, env = "GIGASTT_WORD_TIMESTAMPS", default_value_t = false)]
-    word_timestamps: bool,
 }
 
 #[allow(clippy::large_enum_variant)]
@@ -441,106 +305,6 @@ enum Commands {
     },
 }
 
-#[allow(clippy::too_many_arguments)]
-fn build_limits(
-    config_path: Option<&str>,
-    idle_timeout_secs: Option<u64>,
-    ws_frame_max_bytes: Option<usize>,
-    body_limit_bytes: Option<usize>,
-    rate_limit_per_minute: Option<u32>,
-    rate_limit_burst: Option<u32>,
-    max_session_secs: Option<u64>,
-    shutdown_drain_secs: Option<u64>,
-    pool_checkout_timeout_secs: Option<u64>,
-    inference_timeout_secs: Option<u64>,
-    jobs_enabled: Option<bool>,
-    jobs_ttl_secs: Option<u64>,
-    jobs_max: Option<usize>,
-    jobs_max_bytes: Option<usize>,
-    jobs_retry: Option<u32>,
-) -> anyhow::Result<RuntimeLimits> {
-    let mut limits = if let Some(path) = config_path {
-        server::config::load_config_file(std::path::Path::new(path))?
-    } else {
-        RuntimeLimits::default()
-    };
-    if let Some(v) = idle_timeout_secs {
-        limits.idle_timeout_secs = v;
-    }
-    if let Some(v) = ws_frame_max_bytes {
-        limits.ws_frame_max_bytes = v;
-    }
-    if let Some(v) = body_limit_bytes {
-        limits.body_limit_bytes = v;
-    }
-    if let Some(v) = rate_limit_per_minute {
-        limits.rate_limit_per_minute = v;
-    }
-    if let Some(v) = rate_limit_burst {
-        limits.rate_limit_burst = v;
-    }
-    if limits.rate_limit_per_minute > 0 && limits.rate_limit_burst == 0 {
-        anyhow::bail!("--rate-limit-burst must be > 0 when --rate-limit-per-minute is enabled");
-    }
-    if let Some(v) = max_session_secs {
-        limits.max_session_secs = v;
-    }
-    if let Some(v) = shutdown_drain_secs {
-        limits.shutdown_drain_secs = v;
-    }
-    if let Some(v) = pool_checkout_timeout_secs {
-        limits.pool_checkout_timeout_secs = v;
-    }
-    if let Some(v) = inference_timeout_secs {
-        limits.inference_timeout_secs = v;
-    }
-    if let Some(v) = jobs_enabled {
-        limits.jobs_enabled = v;
-    }
-    if let Some(v) = jobs_ttl_secs {
-        limits.jobs_ttl_secs = v;
-    }
-    if let Some(v) = jobs_max {
-        limits.jobs_max = v;
-    }
-    if let Some(v) = jobs_max_bytes {
-        limits.jobs_max_bytes = v;
-    }
-    if let Some(v) = jobs_retry {
-        limits.jobs_retry = v;
-    }
-    Ok(limits)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn build_server_config(
-    port: u16,
-    host: String,
-    allow_origin: Vec<String>,
-    cors_allow_any: bool,
-    limits: RuntimeLimits,
-    metrics: bool,
-    metrics_listen: std::net::SocketAddr,
-    trust_proxy: bool,
-    config: Option<String>,
-    batch_pool_size: usize,
-) -> ServerConfig {
-    ServerConfig {
-        port,
-        host,
-        origin_policy: OriginPolicy {
-            allow_any: cors_allow_any,
-            allowed_origins: allow_origin,
-        },
-        limits,
-        metrics_enabled: metrics,
-        metrics_listen,
-        trust_proxy,
-        config_path: config.map(std::path::PathBuf::from),
-        batch_pool_size,
-    }
-}
-
 /// clap value parser for `--model-variant`. Accepts `rnnt` / `e2e_rnnt` /
 /// `ml_ctc` / `ml_ctc_large` (case-insensitive); see [`ModelVariant::from_str`].
 fn parse_model_variant(s: &str) -> Result<ModelVariant, String> {
@@ -550,457 +314,6 @@ fn parse_model_variant(s: &str) -> Result<ModelVariant, String> {
 /// Parse the `download --progress` value (`human` | `json`).
 fn parse_progress_mode(s: &str) -> Result<ProgressMode, String> {
     s.parse()
-}
-
-/// Build the synchronous transcribe closure injected into the batch / watch
-/// runners: check out a pool triplet (blocking — the runner calls it from a
-/// blocking thread) and transcribe one file.
-fn make_transcribe_fn(engine: std::sync::Arc<inference::Engine>) -> batch::TranscribeFn {
-    std::sync::Arc::new(move |path: std::path::PathBuf| {
-        let path_str = path
-            .to_str()
-            .ok_or_else(|| anyhow::anyhow!("non-UTF-8 path: {}", path.display()))?;
-        let mut guard = engine
-            .pool
-            .checkout_blocking()
-            .map_err(|e| anyhow::anyhow!("session pool closed: {e}"))?;
-        Ok(engine.transcribe_file(path_str, &mut guard)?)
-    })
-}
-
-/// Cancellation token fired on SIGINT, shared by the batch / watch runners for
-/// graceful shutdown (finish in-flight files, stop scheduling new ones).
-fn ctrl_c_token() -> tokio_util::sync::CancellationToken {
-    let token = tokio_util::sync::CancellationToken::new();
-    let inner = token.clone();
-    tokio::spawn(async move {
-        match tokio::signal::ctrl_c().await {
-            Ok(()) => inner.cancel(),
-            Err(e) => tracing::warn!("Failed to listen for Ctrl-C: {e}"),
-        }
-    });
-    token
-}
-
-/// Build the [`batch::BatchOptions`] shared by `transcribe-batch` and `watch`
-/// from the parsed CLI flags.
-fn build_batch_options(
-    input_dir: &str,
-    output_dir: &str,
-    pool_size: usize,
-    retries: u32,
-    out: &BatchOutputArgs,
-) -> anyhow::Result<batch::BatchOptions> {
-    Ok(batch::BatchOptions {
-        input_dir: std::path::PathBuf::from(input_dir),
-        output_dir: std::path::PathBuf::from(output_dir),
-        formats: batch::parse_formats(&out.format).map_err(|e| anyhow::anyhow!("{e}"))?,
-        render_opts: RenderOpts {
-            max_chars_per_line: out.max_chars_per_line.unwrap_or(80),
-            max_words_per_line: out.max_words_per_line.unwrap_or(14),
-            include_word_timestamps: out.word_timestamps,
-        },
-        move_to: out.move_to.as_deref().map(std::path::PathBuf::from),
-        delete_source: out.delete_source,
-        concurrency: pool_size,
-        retries,
-    })
-}
-
-/// Packaging: rebuild INT8 encoder from a local FP32 ONNX (no download).
-fn run_quantize(model_dir: String, force: bool) -> anyhow::Result<()> {
-    // Order matters for lean INT8 installs: if INT8 is already present
-    // and `--force` is off, no-op without looking for FP32 (CI/e2e and
-    // operators who only have the prequantized bundle).
-    let dir = std::path::Path::new(&model_dir);
-    let resolved = model::ModelVariant::detect_in_dir(dir).unwrap_or_default();
-    let input = dir.join(resolved.encoder_file());
-    let output = dir.join(resolved.encoder_int8_file());
-    if output.exists() && !force {
-        tracing::info!("INT8 model already exists: {}", output.display());
-        tracing::info!("Use --force to re-quantize.");
-        return Ok(());
-    }
-    if !input.is_file() {
-        anyhow::bail!(
-            "FP32 encoder not found at {} — `quantize` needs the FP32 ONNX as packaging source. \
-             For runtime, use `gigastt download` (lean INT8 only).",
-            input.display()
-        );
-    }
-    gigastt_core::quantize::quantize_model(&input, &output)?;
-    tracing::info!("Quantized model saved to {}", output.display());
-    Ok(())
-}
-
-/// Prune optimized/CoreML caches and optionally content-hash-dedupe the model dir.
-fn run_cache_gc(model_dir: String, dry_run: bool, dedupe: bool) -> anyhow::Result<()> {
-    let dir = std::path::Path::new(&model_dir);
-    let prune = model::prune_optimized_cache(dir, dry_run)?;
-    let action = if dry_run { "would free" } else { "freed" };
-    println!(
-        "optimized_cache: kept {} graph(s), removed {} ({} {:.1} MiB)",
-        prune.kept.len(),
-        prune.removed.len(),
-        action,
-        prune.freed_bytes as f64 / (1024.0 * 1024.0),
-    );
-    for p in &prune.removed {
-        println!("  - {}", p.display());
-    }
-    let coreml = model::prune_coreml_cache(dir, dry_run)?;
-    println!(
-        "coreml_cache: kept {}, removed {} stale ({} {:.1} MiB)",
-        if coreml.kept.is_some() {
-            "current"
-        } else {
-            "none"
-        },
-        coreml.removed.len(),
-        action,
-        coreml.freed_bytes as f64 / (1024.0 * 1024.0),
-    );
-    for p in &coreml.removed {
-        println!("  - {}", p.display());
-    }
-    if dedupe {
-        let d = model::dedupe_model_dir(dir, dry_run)?;
-        println!(
-            "dedupe: {} group(s), {} hardlink(s), {} {:.1} MiB",
-            d.groups,
-            d.hardlinked,
-            action,
-            d.freed_bytes as f64 / (1024.0 * 1024.0),
-        );
-    }
-    Ok(())
-}
-
-/// Download the lean INT8 model (and optional side assets). On failure or
-/// SIGINT exits the process with a progress-kind exit code (never returns `Err`
-/// for those paths).
-async fn run_download(
-    model_dir: String,
-    model_variant: ModelVariant,
-    #[cfg(feature = "diarization")] skip_diarization: bool,
-    progress: ProgressMode,
-    #[cfg(feature = "ane")] ane: bool,
-) -> anyhow::Result<()> {
-    model::set_progress_mode(progress);
-    // `download` is an explicit action: the requested variant maps to
-    // the default (Rnnt) so a bare `gigastt download` fetches something
-    // useful. **INT8 only** — lean prequantized bundle (or CTC INT8 from HF).
-    //
-    // The flow runs on its own task: large-file SHA-256 verify is
-    // synchronous, and polled inline it would starve the select's signal
-    // branch — Ctrl-C must interrupt immediately (sidecar cancel path).
-    let dl_model_dir = model_dir.clone();
-    let mut download = tokio::spawn(async move {
-        let model_dir = dl_model_dir;
-        if model_variant.is_ctc() {
-            // CTC heads: HF pre-quantized INT8 encoder (+ vocab) directly.
-            model::ensure_model_variant(Some(model_variant), &model_dir).await?;
-        } else {
-            // RNN-T: lean INT8 bundle from the pinned Release.
-            model::ensure_prequantized_model_variant(Some(model_variant), &model_dir).await?;
-        }
-        #[cfg(feature = "diarization")]
-        {
-            if !skip_diarization {
-                model::ensure_speaker_model(&model_dir).await?;
-            }
-        }
-        #[cfg(feature = "ane")]
-        if ane {
-            let ane_dir = model::default_ane_model_dir();
-            model::ensure_ane_packages(&ane_dir).await?;
-            tracing::info!("ANE encoder packages ready at {ane_dir}");
-        }
-        tracing::info!("Model ready at {model_dir}");
-        anyhow::Ok(())
-    });
-    // Resolves only on a *delivered* SIGINT. A failed handler
-    // registration is logged and parks forever — it must not fabricate
-    // an interrupt and abort a healthy download with exit 130.
-    let interrupted = async {
-        match tokio::signal::ctrl_c().await {
-            Ok(()) => (),
-            Err(e) => {
-                tracing::warn!("Failed to listen for Ctrl-C: {e}");
-                std::future::pending::<()>().await
-            }
-        }
-    };
-    tokio::select! {
-        joined = &mut download => {
-            // Flatten the JoinHandle: a panic inside the download task
-            // is reported through the same error contract.
-            let result = joined
-                .map_err(|e| anyhow::anyhow!("download task failed: {e}"))
-                .and_then(|r| r);
-            match result {
-                Ok(()) => {
-                    model::emit_progress_event(&model::ProgressEvent::Done {
-                        model_dir: model_dir.clone(),
-                    });
-                }
-                Err(e) => {
-                    let kind = model::classify_download_error(&e);
-                    model::emit_progress_event(&model::ProgressEvent::Error {
-                        kind,
-                        message: format!("{e:#}"),
-                    });
-                    // Same rendering anyhow's `Termination` would print,
-                    // then the documented per-kind exit code (all != 0).
-                    eprintln!("Error: {e:?}");
-                    std::process::exit(kind.exit_code());
-                }
-            }
-        }
-        _ = interrupted => {
-            model::emit_progress_event(&model::ProgressEvent::Error {
-                kind: model::ProgressErrorKind::Interrupted,
-                message: "interrupted by SIGINT".to_string(),
-            });
-            eprintln!("Interrupted by Ctrl-C");
-            std::process::exit(model::ProgressErrorKind::Interrupted.exit_code());
-        }
-    }
-    Ok(())
-}
-
-/// Offline single-file transcription (telephony / stereo / mono).
-#[allow(clippy::too_many_arguments)]
-async fn run_transcribe(
-    file: String,
-    model_dir: String,
-    model_variant: Option<ModelVariant>,
-    punctuation: PunctuationMode,
-    punct_model_dir: String,
-    itn: ItnMode,
-    hotwords_file: Option<String>,
-    hotwords_default: bool,
-    hotwords_boost: Option<f32>,
-    vad: bool,
-    vad_threshold: Option<f32>,
-    vad_min_silence_ms: Option<u32>,
-    vad_model_dir: String,
-    encoder_intra_threads: Option<usize>,
-    format: String,
-    output: Option<String>,
-    max_chars_per_line: Option<usize>,
-    max_words_per_line: Option<usize>,
-    word_timestamps: bool,
-    stereo_speakers: bool,
-    codec: Option<String>,
-    sample_rate: Option<u32>,
-) -> anyhow::Result<()> {
-    // Single-triplet pool for offline file transcription; when the
-    // thread count is unset it defaults to every logical CPU (one
-    // running triplet), else the explicit value is used as-is.
-    let engine = EngineRecipe::offline(
-        model_dir,
-        model_variant,
-        punctuation,
-        punct_model_dir,
-        itn,
-        hotwords_file,
-        hotwords_default,
-        hotwords_boost,
-        vad,
-        vad_threshold,
-        vad_min_silence_ms,
-        vad_model_dir,
-        encoder_intra_threads,
-        1,
-    )
-    .load_offline_engine()
-    .await?;
-    let mut guard = engine.pool.checkout().await?;
-    let result = if let Some(codec_name) = codec.as_deref() {
-        // Raw headerless telephony input: decode via the codec tables
-        // straight to mono 16 kHz f32 and hand the samples to the engine.
-        let telephony_codec =
-            inference::audio::TelephonyCodec::from_name(codec_name).ok_or_else(|| {
-                anyhow::anyhow!("unsupported codec '{codec_name}' (supported: pcmu, pcma, g722)")
-            })?;
-        // clap enforces `--sample-rate` when `--codec` is given; keep a
-        // graceful error instead of an unwrap in case that ever changes.
-        let rate =
-            sample_rate.ok_or_else(|| anyhow::anyhow!("--sample-rate is required with --codec"))?;
-        telephony_codec
-            .validate_sample_rate(rate)
-            .map_err(|e| anyhow::anyhow!("{e}"))?;
-        let raw =
-            std::fs::read(&file).with_context(|| format!("Failed to open audio file: {file}"))?;
-        let mut samples = inference::audio::decode_telephony_raw(&raw, telephony_codec, rate)?;
-        // NOT a no-op: preserve transcript byte-identity. The former path
-        // encoded these samples into an in-memory WAV and let the engine
-        // decode it back, which snapped every value to 16-bit PCM. Passing
-        // the raw f32 straight through would change the transcript (measured:
-        // 153 token edits on a 9-minute call), so reproduce that quantization
-        // in place — clamp to [-1, 1], round via 32767, normalise via 32768,
-        // matching `encode_wav_pcm16` and the PCM decode path. Whether
-        // full-precision f32 is better is an open WER question tracked in
-        // roadmap/ (telephony precision); do not silently drop this snap.
-        for s in samples.iter_mut() {
-            let v = if s.is_finite() {
-                s.clamp(-1.0, 1.0)
-            } else {
-                0.0
-            };
-            *s = (v * 32767.0).round() as i16 as f32 / 32768.0;
-        }
-        engine.transcribe_request(
-            inference::TranscribeRequest::new(inference::TranscribeSource::Samples(&samples)),
-            &mut guard,
-        )
-    } else if stereo_speakers {
-        let channels = inference::audio::load_audio_channels(&file)?;
-        let fallback_reason = match channels.len() {
-            0 => Some("no channels"),
-            1 => Some("mono audio"),
-            2 if inference::audio::is_dual_mono(&channels) => Some("dual-mono audio"),
-            n if n > 2 => Some("more than two channels"),
-            _ => None,
-        };
-        if let Some(reason) = fallback_reason {
-            tracing::warn!(
-                "--stereo-speakers requested but {reason} detected; falling back to mono transcription"
-            );
-            engine.transcribe_file(&file, &mut guard)
-        } else {
-            engine.transcribe_channels(&channels, &mut guard)
-        }
-    } else {
-        engine.transcribe_file(&file, &mut guard)
-    };
-    drop(guard);
-    let result = result?;
-
-    let format = ExportFormat::from_str(&format).map_err(|e| anyhow::anyhow!("{e}"))?;
-    let opts = RenderOpts {
-        max_chars_per_line: max_chars_per_line.unwrap_or(80),
-        max_words_per_line: max_words_per_line.unwrap_or(14),
-        include_word_timestamps: word_timestamps,
-    };
-    let rendered = format.render(&result, &opts);
-
-    match output {
-        Some(path) => {
-            std::fs::write(&path, rendered).with_context(|| format!("failed to write {path}"))?;
-            tracing::info!("Wrote {} export to {path}", format);
-        }
-        None => println!("{rendered}"),
-    }
-    Ok(())
-}
-
-async fn run_transcribe_batch(
-    input_dir: String,
-    output_dir: String,
-    eng: OfflineEngineArgs,
-    out: BatchOutputArgs,
-) -> anyhow::Result<()> {
-    let engine = EngineRecipe::offline(
-        eng.model_dir,
-        eng.model_variant,
-        eng.punctuation,
-        eng.punct_model_dir,
-        eng.itn,
-        eng.hotwords_file,
-        eng.hotwords_default,
-        eng.hotwords_boost,
-        eng.vad,
-        eng.vad_threshold,
-        eng.vad_min_silence_ms,
-        eng.vad_model_dir,
-        eng.encoder_intra_threads,
-        eng.pool_size,
-    )
-    .load_offline_engine()
-    .await?;
-    let opts = build_batch_options(
-        &input_dir,
-        &output_dir,
-        eng.pool_size,
-        out.retries.unwrap_or(0),
-        &out,
-    )?;
-    let summary = batch::run_batch(
-        &opts,
-        make_transcribe_fn(std::sync::Arc::new(engine)),
-        ctrl_c_token(),
-    )
-    .await?;
-    tracing::info!(
-        processed = summary.processed,
-        failed = summary.failed,
-        skipped = summary.skipped,
-        "batch finished"
-    );
-    if summary.interrupted {
-        // Same contract as `download`: SIGINT exits 130.
-        std::process::exit(130);
-    }
-    if summary.failed > 0 {
-        std::process::exit(1);
-    }
-    Ok(())
-}
-
-async fn run_watch(
-    input_dir: String,
-    output_dir: String,
-    eng: OfflineEngineArgs,
-    out: BatchOutputArgs,
-    poll_interval_ms: u64,
-    settle_polls: u32,
-) -> anyhow::Result<()> {
-    let engine = EngineRecipe::offline(
-        eng.model_dir,
-        eng.model_variant,
-        eng.punctuation,
-        eng.punct_model_dir,
-        eng.itn,
-        eng.hotwords_file,
-        eng.hotwords_default,
-        eng.hotwords_boost,
-        eng.vad,
-        eng.vad_threshold,
-        eng.vad_min_silence_ms,
-        eng.vad_model_dir,
-        eng.encoder_intra_threads,
-        eng.pool_size,
-    )
-    .load_offline_engine()
-    .await?;
-    let opts = batch::WatchOptions {
-        batch: build_batch_options(
-            &input_dir,
-            &output_dir,
-            eng.pool_size,
-            out.retries.unwrap_or(2),
-            &out,
-        )?,
-        poll_interval: std::time::Duration::from_millis(poll_interval_ms),
-        settle_polls,
-    };
-    let summary = batch::run_watch(
-        &opts,
-        make_transcribe_fn(std::sync::Arc::new(engine)),
-        ctrl_c_token(),
-    )
-    .await?;
-    tracing::info!(
-        processed = summary.processed,
-        failed = summary.failed,
-        "watch stopped"
-    );
-    if summary.failed > 0 {
-        std::process::exit(1);
-    }
-    Ok(())
 }
 
 #[tokio::main]
@@ -1153,9 +466,12 @@ async fn main() -> anyhow::Result<()> {
 #[cfg(test)]
 mod tests {
     use super::serve::{
-        ServeProfile, ensure_bind_allowed, ensure_metrics_bind_allowed, is_loopback_host,
+        ServeProfile, build_limits, build_server_config, ensure_bind_allowed,
+        ensure_metrics_bind_allowed, is_loopback_host,
     };
     use super::*;
+    use gigastt::server;
+    use gigastt::server::RuntimeLimits;
 
     // Serialize tests that mutate process env vars to avoid races under
     // cargo test's default multi-threaded harness (used by tarpaulin).
